created : separate pages and data for each club
