+++ pages/ieee_comsoc/events.js
@@ -2,10 +2,9 @@ import React from 'react';
 import {Upcoming, Past} from '../../components/events';
 import Navigation from '../../components/navbar';
 import Head from 'next/head';
-import upcoming_events from '../../public/data/ieee_cs/upcoming_events.json';
-import past_events from '../../public/data/ieee_cs/past_events.json';
+import upcoming_events from '../../public/data/ieee_comsoc/upcoming_events.json';
+import past_events from '../../public/data/ieee_comsoc/past_events.json';
 import {Nav, NavAcross} from "../../components/nav";
-
 
 const Events = () => {
     
@@ -15,7 +14,7 @@ const Events = () => {
         <title>IEEE_CS</title>
       </Head>
       <NavAcross></NavAcross>
-      <Navigation team_home="/ieee_cs" events="/ieee_cs/events" gallery="/ieee_cs/gallery"></Navigation>
+      <Navigation team_home="/ieee_comsoc" events="/ieee_comsoc/events" gallery="/ieee_comsoc/gallery"></Navigation>
       <Upcoming data={upcoming_events}></Upcoming>
       <Past data={past_events}></Past>
     </>
