+++ pages/ieee_cs/gallery.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+function Gallery(){
+    return (
+        <h1>This page is under construction</h1>
+    )
+}
+
+export default Gallery;
