+++ public/assets/team/ieee_cs/heads/Vice_Chair.jpg
