+++ public/assets/team/ieee_cs/other_members/Harshida S P_CoreTeam.JPG
