+++ public/assets/team/ieee_cs/other_members/Sanjai Balajee core team.png
