+++ public/assets/team/ieee_cs/other_members/Sanjjit S_CoreTeam.jpg
