+++ public/assets/team/ieee_cs/other_members/Srujana_Design.PNG
