+++ public/assets/team/ieee_cs/other_members/Sushmithaa _ core team .jpg
