+++ public/assets/team/ieee_cs/other_members/V.S.Pranav_CoreTeam.jpeg
