+++ public/data/ieee_comsoc/core_members.json
@@ -1,0 +1,29 @@
+[
+    {
+        "name": "Anirudh A",
+        "image": "/assets/team/ieee_comsoc/heads/Treasurer.jpg",
+        "post": "Treasurer",
+        "contact": {
+            "email": { 
+                "has": false, 
+                "url": null
+            },
+
+            "linkedin": {
+                "has": false,
+                "url": null
+            },
+            "github": {
+                "has": false,
+                "url": null
+            },
+
+            "instagram": {
+                "has": false,
+                "url": null
+            }
+    
+        }
+    }
+    
+]
