+++ public/data/ieee_comsoc/past_events.json
@@ -1,0 +1,10 @@
+[
+    {
+        "title" : "dfdf",
+        "description" : "empty desc",
+        "date" : "empyt date",
+        "venue" : "chennai",
+        "image" : "/assets/team/ieee_comsoc/events/Tect.jpeg"
+    }
+
+]
