+++ public/data/ieee_cs/upcoming_events.json
@@ -4,7 +4,7 @@
         "description" : "Brace yourselves as we are bringing you our first offline event, TEC-T. Come and learn about this digital revolution from the speaker(Mr. Mohamed Rawidean), who has 20 years of experience in the design, development, and installation of smart-home systems.",
         "date" : "Oct 11",
         "venue" : "Central Seminar Hall (ECE)",
-        "image" : "/assets/team/IEEE_CS/Events/Tect.jpeg"
+        "image" : "/assets/team/ieee_cs/events/Tect.jpeg"
     }
