+++ public/data/ieee_embs/past_events.json
@@ -1,0 +1,10 @@
+[
+    {
+        "title" : "Past event 2",
+        "description" : "dummy desc",
+        "date" : "dummy date",
+        "venue" : "chennai2",
+        "image" : "/assets/team/ieee_embs/events/Tect.jpeg"
+    }
+    
+]
