+++ public/data/ieee_embs/upcoming_events.json
@@ -1,0 +1,11 @@
+[
+    {
+        "title": "Upcoming event 1",
+        "description" : "dfsdfs",
+        "date" : "Oct 11",
+        "venue" : "SSN RWRW",
+        "image" : "/assets/team/ieee_embs/events/Tect.jpeg"
+    }
+
+    
+]
