+++ public/data/ieee_photonics/past_events.json
@@ -1,0 +1,9 @@
+[
+    {
+        "title" : "Past Event 10",
+        "description" : "Description here",
+        "date" : "Nov 19",
+        "venue" : "WOrld",
+        "image" : "/assets/team/ieee_photonics/events/Tect.jpeg"
+    }
+]
