+++ public/data/ieee_photonics/upcoming_events.json
@@ -1,0 +1,11 @@
+[
+    {
+        "title": "Upcoming event 10",
+        "description" : "this is description",
+        "date" : "Jan 01",
+        "venue" : "SSN Audi",
+        "image" : "/assets/team/ieee_photonics/events/Tect.jpeg"
+    }
+
+    
+]
